Support multiple server codecs
This allows a server to provide multiple codecs that it supports and we will find the best translation path to reach any of them. It also means that
a client with the same version as any of the supported codecs will also be short circuited.

Squashed commit of the following:

commit 74b3ba404cd70d07dd9c37fc4b0f638ccc2f90e7
Author: bundabrg <bundabrg@grieve.com.au>
Date:   Wed Dec 9 10:26:02 2020 +0800

    Clean up

commit 7a5c54b2e7acb31efa4f36339b31c2aa34945d7c
Author: bundabrg <bundabrg@grieve.com.au>
Date:   Wed Dec 9 10:04:51 2020 +0800

    First look at multiple toCodecs