Fix buckets for Education
When interacticing holding an item we will check if the client interacts with both a block and air. If only 1, we send through the block and block_interact packet, otherwise we suppress the block_interact packet and send the other packets. Solves buckets and opening chests whilst holding a bucket.